Streamline function pattern
Function pattern is now a much more general regular expression that is not syntax aware. The advantage is that this is far more robust and faster. The disadvantage is that it no longer can help catch syntax errors. However, this is of minimal value since go tools are typically run and will also provide the syntax errors.

Potentially could revisit the regex patter for the full syntax but a better solution would be to move all this to a native BBLM that uses the Go ast or a real parser grammer to check it.

This seems like a solid middle ground that can avoid the edge cases and will work much faster with less resources.