Update README for project name change
Changed the name of the project from Go.bbpackage to Go-bbpackage. The reason
for the change is that this package needs to be built and using `Go.bbpackage`
implies it can be directly dropped into your BBEdit application support directory.
This is not true so the name change is to make it clear that you need to check it out
and build it.

The project will be refactored to generate a `Go.bbpackage` artifact and directly
put into the support directory.